Use zeitwerk for loading and remove unused gems (#2)
* Use zeitwerk for loading and remove unused gems

* Dump version
