+++ Gemfile
@@ -1,4 +1,3 @@
 source "https://rubygems.org"
 
-gem "codeclimate-test-reporter", group: :test, require: nil
 gemspec
