+++ lib/norma43.rb
@@ -1,8 +1,12 @@
-require "norma43/version"
-require "norma43/parser"
+require "zeitwerk"
+
+loader = Zeitwerk::Loader.for_gem
+loader.setup # ready!
 
 module Norma43
   def self.parse text
     Parser.new(text).result
   end
 end
+
+loader.eager_load
