+++ lib/norma43/line_parsers/additional_item.rb
@@ -1,0 +1,9 @@
+module Norma43
+  module LineParsers
+    class AdditionalItem < LineParser
+      field :data_code, 2..3, :integer
+      field :item_1, 4..41
+      field :item_2, 42..79
+    end
+  end
+end
