+++ lib/norma43/line_parsers/document_end.rb
@@ -1,0 +1,7 @@
+module Norma43
+  module LineParsers
+    class DocumentEnd < LineParser
+      field :record_number, 20..25, :integer
+    end
+  end
+end
