+++ lib/norma43/line_parsers/line_parser.rb
@@ -1,42 +1,42 @@
-require "norma43/utils/typecaster"
-
 module Norma43
-  class LineParser
-    attr_reader :line
-    def initialize line
-      @line = line
-    end
-
-    def attributes
-      self.class.field_names.each_with_object({}) do |field, attrs|
-        attrs[field] = self.public_send(field)
+  module LineParsers
+    class LineParser
+      attr_reader :line
+      def initialize line
+        @line = line
       end
-    end
 
-    def self.field name, range, type = :string
-      self.field_names.push name
-
-      define_method name do
-        if range.is_a?(Array) # let multivalued attribute
-          range.map { |r| value_at_position(r, type) }.compact
-        else
-          value_at_position range, type
+      def attributes
+        self.class.field_names.each_with_object({}) do |field, attrs|
+          attrs[field] = self.public_send(field)
         end
       end
-    end
 
-    private
+      def self.field name, range, type = :string
+        self.field_names.push name
 
-    def self.field_names
-      @field_names ||= []
-    end
+        define_method name do
+          if range.is_a?(Array) # let multivalued attribute
+            range.map { |r| value_at_position(r, type) }.compact
+          else
+            value_at_position range, type
+          end
+        end
+      end
 
-    def value_at_position range, type
-      typecast line[range].to_s.strip, type
-    end
+      private
 
-    def typecast value, type
-      Typecaster.cast value, type
+      def self.field_names
+        @field_names ||= []
+      end
+
+      def value_at_position range, type
+        typecast line[range].to_s.strip, type
+      end
+
+      def typecast value, type
+        Norma43::Utils::Typecaster.cast value, type
+      end
     end
   end
 end
