+++ lib/norma43/utils/string_helpers.rb
@@ -1,10 +1,12 @@
 module Norma43
-  module StringHelpers
-    def self.underscore word
-      word.gsub!(/([A-Z\d]+)([A-Z][a-z])/,'\1_\2')
-      word.gsub!(/([a-z\d])([A-Z])/,'\1_\2')
-      word.tr!("-", "_")
-      word.downcase
+  module Utils
+    module StringHelpers
+      def self.underscore word
+        word.gsub!(/([A-Z\d]+)([A-Z][a-z])/,'\1_\2')
+        word.gsub!(/([a-z\d])([A-Z])/,'\1_\2')
+        word.tr!("-", "_")
+        word.downcase
+      end
     end
   end
 end
