+++ lib/norma43/version.rb
@@ -1,3 +1,3 @@
 module Norma43
-  VERSION = "1.0.1"
+  VERSION = "2.0.0"
 end
