+++ norma43_parser.gemspec
@@ -1,4 +1,3 @@
-# coding: utf-8
 lib = File.expand_path('../lib', __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
 require "norma43/version"
@@ -12,16 +11,17 @@ Gem::Specification.new do |spec|
   spec.homepage      = "https://github.com/sequra/norma43_parser"
   spec.license       = "MIT"
 
-  spec.required_ruby_version = "~> 2.0"
+  spec.required_ruby_version = "~> 2.4"
 
   spec.files         = `git ls-files -z`.split("\x0")
   spec.executables   = spec.files.grep(%r{^bin/}) { |f| File.basename(f) }
   spec.test_files    = spec.files.grep(%r{^(test|spec|features)/})
   spec.require_paths = ["lib"]
 
-  spec.add_development_dependency "bundler", "~> 1.5"
-  spec.add_development_dependency "rake"
-  spec.add_development_dependency "byebug"
-  spec.add_development_dependency "rspec"
-  spec.add_runtime_dependency "virtus"
+  spec.add_runtime_dependency "virtus",   "~> 1.0"
+  spec.add_runtime_dependency "zeitwerk", "~> 2.0"
+
+  spec.add_development_dependency "bundler", "~> 1.0"
+  spec.add_development_dependency "rake",    "~> 10.0"
+  spec.add_development_dependency "rspec",   "~> 3.0"
 end
