+++ spec/norma43/line_parsers/document_end_spec.rb
@@ -1,5 +1,3 @@
-require "norma43/line_parsers/line_parsers"
-
 module Norma43
   module LineParsers
     RSpec.describe DocumentEnd do
