+++ spec/norma43/line_processors/account_end_spec.rb
@@ -1,19 +1,16 @@
-require "norma43/utils/contexts"
-require "norma43/line_processors"
-
 module Norma43
   module LineProcessors
-    RSpec.describe "AccountEnd" do
+    RSpec.describe AccountEnd do
       let :line do
         double "Line", attributes: {}
       end
 
-      let(:account){ Models::Account.new }
-      let(:contexts){ Contexts.new(
+      let(:account){ Norma43::Models::Account.new }
+      let(:contexts){ Norma43::Utils::Contexts.new(
         [
-          Models::Document.new,
+          Norma43::Models::Document.new,
           account,
-          Models::Transaction.new
+          Norma43::Models::Transaction.new
         ])
       }
 
@@ -31,4 +28,3 @@ module Norma43
     end
   end
 end
-
