+++ spec/norma43/line_processors/transaction_spec.rb
@@ -1,16 +1,13 @@
-require "norma43/utils/contexts"
-require "norma43/line_processors"
-
 module Norma43
   module LineProcessors
     RSpec.describe "Transaction" do
       let(:line){ double "Line", attributes: {} }
-      let(:account){ Models::Account.new }
-      let(:contexts){ Contexts.new(
+      let(:account){ Norma43::Models::Account.new }
+      let(:contexts){ Norma43::Utils::Contexts.new(
         [
-          Models::Document.new ,
+          Norma43::Models::Document.new ,
           account,
-          Models::Transaction.new
+          Norma43::Models::Transaction.new
         ])
       }
