+++ spec/spec_helper.rb
@@ -1,8 +1,3 @@
-if ENV["CI"]
-  require "codeclimate-test-reporter"
-  CodeClimate::TestReporter.start
-end
-
 RSpec.configure do |config|
   config.order = :random
   Kernel.srand config.seed
